optimize:modify auto config logic
