+++ seata-spring-autoconfigure/seata-spring-autoconfigure-core/src/main/java/org/apache/seata/spring/boot/autoconfigure/StarterConstants.java
@@ -80,7 +80,6 @@ public interface StarterConstants {
     String SERVER_HTTP_PREFIX = SERVER_PREFIX + ".http";
 
     String SERVER_HTTP_FILTER_PREFIX = SERVER_HTTP_PREFIX + ".filter";
-    String SERVER_HTTP_FILTER_XSS_PREFIX = SERVER_HTTP_FILTER_PREFIX + ".xss";
 
     String METRICS_PREFIX = SEATA_PREFIX + ".metrics";
